Add inputHandler.js for mouse and keyboard
Added the `Input` object to keep track of the current state of all mouse and key inputs, touch inputs to be tracked later.
The current state of user input can now be used anywhere in the program, not just in an event callback. Example of utility is spawning an object for every frame that `Input.mouse.isHeld` is true.